requirements.txt
# WiFi Tester Pro v6.0 - Dependencies
# Install with: pip install -r requirements.txt

# GUI Framework
customtkinter>=5.2.0
Pillow>=10.0.0

# System & Network
psutil>=5.9.0
requests>=2.31.0

# Packet manipulation (Kali Linux only - optional on Windows)
scapy>=2.5.0

# Plotting (optional - for signal graphs)
matplotlib>=3.7.0
